+++ EnvironmentSimulator/Modules/Controllers/ControllerACC.cpp
@@ -123,17 +123,13 @@ void ControllerACC::Step(double timeStep)
             double dHeading          = GetAbsAngleDifference(object_->pos_.GetH(), pivot_obj->pos_.GetH());
             if (dHeading < M_PI_2)  // objects are pointing roughly in the same direction
             {
-                adjustedGapLength -=
-                    (static_cast<double>(object_->boundingbox_.dimensions_.length_) / 2.0 + static_cast<double>(object_->boundingbox_.center_.x_)) +
-                    (static_cast<double>(pivot_obj->boundingbox_.dimensions_.length_) / 2.0 -
-                     static_cast<double>(pivot_obj->boundingbox_.center_.x_));
+                adjustedGapLength -= (object_->boundingbox_.dimensions_.length_ / 2.0 + object_->boundingbox_.center_.x_) +
+                                     (pivot_obj->boundingbox_.dimensions_.length_ / 2.0 - pivot_obj->boundingbox_.center_.x_);
             }
             else  // objects are pointing roughly in the opposite direction
             {
-                adjustedGapLength -=
-                    (static_cast<double>(object_->boundingbox_.dimensions_.length_) / 2.0 + static_cast<double>(object_->boundingbox_.center_.x_)) +
-                    (static_cast<double>(pivot_obj->boundingbox_.dimensions_.length_) / 2.0 +
-                     static_cast<double>(pivot_obj->boundingbox_.center_.x_));
+                adjustedGapLength -= (object_->boundingbox_.dimensions_.length_ / 2.0 + object_->boundingbox_.center_.x_) +
+                                     (pivot_obj->boundingbox_.dimensions_.length_ / 2.0 + pivot_obj->boundingbox_.center_.x_);
             }
 
             // dLaneId == 0 indicates there is linked path between object lanes, i.e. no lane changes needed
@@ -151,9 +147,7 @@ void ControllerACC::Step(double timeStep)
             double x_local, y_local;
             object_->FreeSpaceDistance(pivot_obj, &y_local, &x_local);
 
-            if (x_local > 0 &&
-                x_local <
-                    1.0 + static_cast<double>(pivot_obj->boundingbox_.dimensions_.length_) + 0.5 * MAX(0.0, currentSpeed_ - pivot_obj->GetSpeed()) &&
+            if (x_local > 0 && x_local < 1.0 + pivot_obj->boundingbox_.dimensions_.length_ + 0.5 * MAX(0.0, currentSpeed_ - pivot_obj->GetSpeed()) &&
                 y_local < 0.2 && y_local > -0.5)  // yield some more for right hand traffic
             {
                 minGapLength = x_local;
