+++ EnvironmentSimulator/Modules/Controllers/ControllerLooming.cpp
@@ -312,17 +312,13 @@ void ControllerLooming::Step(double timeStep)
             double dHeading          = GetAbsAngleDifference(object_->pos_.GetH(), pivot_obj->pos_.GetH());
             if (dHeading < M_PI_2)  // objects are pointing roughly in the same direction
             {
-                adjustedGapLength -=
-                    (static_cast<double>(object_->boundingbox_.dimensions_.length_) / 2.0 + static_cast<double>(object_->boundingbox_.center_.x_)) +
-                    (static_cast<double>(pivot_obj->boundingbox_.dimensions_.length_) / 2.0 -
-                     static_cast<double>(pivot_obj->boundingbox_.center_.x_));
+                adjustedGapLength -= (object_->boundingbox_.dimensions_.length_ / 2.0 + object_->boundingbox_.center_.x_) +
+                                     (pivot_obj->boundingbox_.dimensions_.length_ / 2.0 - pivot_obj->boundingbox_.center_.x_);
             }
             else  // objects are pointing roughly in the opposite direction
             {
-                adjustedGapLength -=
-                    (static_cast<double>(object_->boundingbox_.dimensions_.length_) / 2.0 + static_cast<double>(object_->boundingbox_.center_.x_)) +
-                    (static_cast<double>(pivot_obj->boundingbox_.dimensions_.length_) / 2.0 +
-                     static_cast<double>(pivot_obj->boundingbox_.center_.x_));
+                adjustedGapLength -= (object_->boundingbox_.dimensions_.length_ / 2.0 + object_->boundingbox_.center_.x_) +
+                                     (pivot_obj->boundingbox_.dimensions_.length_ / 2.0 + pivot_obj->boundingbox_.center_.x_);
             }
 
             // dLaneId == 0 indicates there is linked path between object lanes, i.e. no lane changes needed
