+++ EnvironmentSimulator/code-examples/hello_world/hw7_ideal_sensors.cpp
@@ -7,8 +7,8 @@ int main(int, char**)
 {
     SE_Init("../resources/xosc/cut-in.xosc", 0, 1, 0, 0);
 
-    SE_AddObjectSensor(0, 2.0f, 1.0f, 0.5f, 1.57f, 1.0f, 50.0f, 1.57f, MAX_HITS);
-    SE_AddObjectSensor(0, -1.0f, 0.0f, 0.5f, 3.14f, 0.5f, 20.0f, 1.57f, MAX_HITS);
+    SE_AddObjectSensor(0, 2.0, 1.0, 0.5, 1.57, 1.0, 50.0, 1.57, MAX_HITS);
+    SE_AddObjectSensor(0, -1.0, 0.0, 0.5, 3.14, 0.5, 20.0, 1.57, MAX_HITS);
 
     // Turn on visualization of object sensors, toggle key 'r'
     SE_ViewerShowFeature(1, true);
@@ -29,11 +29,7 @@ int main(int, char**)
                 SE_ScenarioObjectState state;
 
                 SE_GetObjectState(objList[k], &state);
-                printf("object[%d] pos: (%.2f, %.2f) heading: %.2f\n",
-                       objList[k],
-                       static_cast<double>(state.x),
-                       static_cast<double>(state.y),
-                       static_cast<double>(state.h));
+                printf("object[%d] pos: (%.2f, %.2f) heading: %.2f\n", objList[k], state.x, state.y, state.h);
             }
         }
     }
